test(mv3): merge mv3 tests into mv2 test files
