+++ test/functional/lib/ipfs-import.test.js
@@ -1,24 +1,9 @@
 'use strict'
-import { describe, it, before, after } from 'mocha'
 import { expect } from 'chai'
-import { useFakeTimers } from 'sinon'
-import browser from 'sinon-chrome'
+import { describe, it } from 'mocha'
 import { formatImportDirectory } from '../../../add-on/src/lib/ipfs-import.js'
 
 describe('ipfs-import.js', function () {
-  before(function () {
-    browser.runtime.id = 'testid'
-    global.document = {}
-    global.browser = browser
-    // ipfs-import depends on webextension/polyfill which can't be imported
-    // in a non-browser environment unless global.browser is stubbed
-
-    // need to force Date to return a particular date
-    global.clock = useFakeTimers({
-      now: new Date(2017, 10, 5, 12, 1, 1)
-    })
-  })
-
   describe('formatImportDirectory', function () {
     it('should change nothing if path is properly formatted and date wildcards are not provided', function () {
       const path = '/ipfs-companion-imports/my-directory/'
@@ -54,11 +39,4 @@ describe('ipfs-import.js', function () {
   //
   // describe('importFiles', function () {
   // })
-
-  after(function () {
-    global.browser.flush()
-    global.clock.restore()
-    delete global.document
-    delete global.browser
-  })
 })
