+++ test/functional/lib/ipfs-request-gateway-recover.test.js
@@ -1,15 +1,15 @@
 'use strict'
-import { describe, it, before, beforeEach, after, afterEach } from 'mocha'
-import sinon from 'sinon'
 import { assert } from 'chai'
-import { URL } from 'url'
+import { after, afterEach, before, beforeEach, describe, it } from 'mocha'
+import sinon from 'sinon'
 import browser from 'sinon-chrome'
-import { initState } from '../../../add-on/src/lib/state.js'
-import createRuntimeChecks from '../../../add-on/src/lib/runtime-checks.js'
-import { createRequestModifier } from '../../../add-on/src/lib/ipfs-request.js'
+import { URL } from 'url'
 import createDnslinkResolver from '../../../add-on/src/lib/dnslink.js'
 import { createIpfsPathValidator } from '../../../add-on/src/lib/ipfs-path.js'
+import { createRequestModifier } from '../../../add-on/src/lib/ipfs-request.js'
 import { optionDefaults } from '../../../add-on/src/lib/options.js'
+import createRuntimeChecks from '../../../add-on/src/lib/runtime-checks.js'
+import { initState } from '../../../add-on/src/lib/state.js'
 
 browser.runtime.id = 'testid'
