+++ test/functional/lib/ipfs-request-workarounds.test.js
@@ -1,21 +1,25 @@
 'use strict'
-import { describe, it, before, beforeEach, after } from 'mocha'
-import { expect, assert } from 'chai'
-import { URL } from 'url' // URL implementation with support for .origin attribute
-import browser from 'sinon-chrome'
-import { initState } from '../../../add-on/src/lib/state.js'
-import createRuntimeChecks from '../../../add-on/src/lib/runtime-checks.js'
-import { createRequestModifier } from '../../../add-on/src/lib/ipfs-request.js'
-import createDNSLinkResolver from '../../../add-on/src/lib/dnslink.js'
-import { createIpfsPathValidator } from '../../../add-on/src/lib/ipfs-path.js'
-import { optionDefaults } from '../../../add-on/src/lib/options.js'
-import { braveNodeType } from '../../../add-on/src/lib/ipfs-client/brave.js'
-import { spoofDnsTxtRecord } from './dnslink.test.js'
+import { assert, expect } from 'chai';
+import { after, before, beforeEach, describe, it } from 'mocha';
+import browser from 'sinon-chrome';
+import { URL } from 'url'; // URL implementation with support for .origin attribute
+import createDNSLinkResolver from '../../../add-on/src/lib/dnslink.js';
+import { braveNodeType } from '../../../add-on/src/lib/ipfs-client/brave.js';
+import { createIpfsPathValidator } from '../../../add-on/src/lib/ipfs-path.js';
+import { createRequestModifier } from '../../../add-on/src/lib/ipfs-request.js';
+import { optionDefaults } from '../../../add-on/src/lib/options.js';
+import createRuntimeChecks from '../../../add-on/src/lib/runtime-checks.js';
+import { initState } from '../../../add-on/src/lib/state.js';
+import isMv3TestingEnabled from '../../helpers/is-mv3-testing-enabled.js';
+import { spoofDnsTxtRecord } from './dnslink.test.js';
 
 describe('modifyRequest processing', function () {
   let state, getState, dnslinkResolver, ipfsPathValidator, modifyRequest, runtime
 
   before(function () {
+    if (isMv3TestingEnabled()) {
+      return this.skip()
+    }
     global.URL = URL
     global.browser = browser
     browser.runtime.id = 'testid'
