+++ test/setup/mocha-setup.js
@@ -1,6 +1,6 @@
 import AbortController from 'abort-controller'
 import { afterEach } from 'mocha'
-import sinon from 'sinon'
+import sinon, { useFakeTimers } from 'sinon'
 import browser from 'sinon-chrome'
 import DeclarativeNetRequestMock from '../functional/lib/redirect-handler/declarativeNetRequest.mock.js'
 import isMv3TestingEnabled from '../helpers/is-mv3-testing-enabled.js'
@@ -14,6 +14,14 @@ global.navigator = {
     writeText: () => {}
   }
 }
+
+global.URL = URL
+browser.tabs = { ...browser.tabs, getCurrent: sinon.stub().resolves({ id: 20 }) }
+
+// need to force Date to return a particular date
+global.clock = useFakeTimers({
+  now: new Date(2017, 10, 5, 12, 1, 1)
+})
 
 if (isMv3TestingEnabled()) {
   const sinonSandbox = sinon.createSandbox()
